requirements.txt
# Time Series Forecasting Project Requirements
# Python 3.10+

# Core ML/Data Science
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Time Series & Statistical Models
statsmodels==0.14.1
pmdarima==2.0.4

# Machine Learning
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.3

# Chronos Model
chronos-forecasting==2.0.0

# PyTorch (for Chronos)
torch==2.1.2
--extra-index-url https://download.pytorch.org/whl/cu118

# Utilities
tqdm==4.66.1
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2

# Jupyter
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1
ipywidgets==8.1.1

# Report generation
jinja2==3.1.2

# Reproducibility
pyarrow==14.0.2  # for parquet support
